Refractor: Shift classes in position module to state module, and remove position module
